[New Feature] Add option for another way of updating widget
For normal situation, using setImageViewUri will automatically scale the picture to the right size for fitting in the widget. But for some devices larger bitmap can be accepted and larger picture means better visual. Unfortunately I don't know how to detect this situation so I made an option for users..

Signed-off-by: Shaka Huang <shaka@cs03.org>